Reorganized project structure and renamed Main.jsx to Home.jsx for clarity
