+++ plane_Catcher2/src/components/Footer/Footer.css
@@ -8,7 +8,7 @@
     font-family: Arial, Helvetica, sans-serif;
 }
 .footer {
-    background: #010223 url("src/assets/imgs/footerBackgroundMap.webp")
+    background: #010223 url("../../assets/imgs/footerBackgroundMap.webp")
     center/cover no-repeat;
 }
