+++ plane_Catcher2/src/components/Footer/Footer.jsx
@@ -1,9 +1,9 @@
-import "../Footer.css";
-import logo from "../assets/imgs/logo.png";
-import whatsapp from "../assets/imgs/whatsapp.png";
-import twitter from "../assets/imgs/twitter.png";
-import instagram from "../assets/imgs/instagram.png";
-import facebook from "../assets/imgs/facebook.png";
+import "./Footer.css";
+import logo from "../../assets/imgs/logo.png";
+import whatsapp from "../../assets/imgs/whatsapp.png";
+import twitter from "../../assets/imgs/twitter.png";
+import instagram from "../../assets/imgs/instagram.png";
+import facebook from "../../assets/imgs/facebook.png";
 
 const Main = () => {
     return (
